alpha beatmaps

+++ CanYouFeelTheRhythm/src/beats/Circle.java
@@ -4,11 +4,11 @@ import java.awt.Color;
 
 public class Circle extends Beat {
 
-	public Circle(String filename, int x, int y, int r, double time) {
+	public Circle(String filename, int x, int y, int r, long time) {
 		super(filename,x,y,r, time);
 	}
 	
-	public Circle(String filename, int x, int y, int r, double time, Color c) {
+	public Circle(String filename, int x, int y, int r, long time, Color c) {
 		super(filename,x,y,r, time);
 		setColor(c);
 	}
