+++ CanYouFeelTheRhythm/src/beats/FallingCircle.java
@@ -5,12 +5,12 @@ import java.awt.Color;
 public class FallingCircle extends Circle {
 
 	
-	public FallingCircle(String filename, int x, int y, int r, double time) {
+	public FallingCircle(String filename, int x, int y, int r, long time) {
 		super(filename, x, y, r, time);
 
 	}
 	
-	public FallingCircle(String filename, int x, int y, int r, double time, Color c) {
+	public FallingCircle(String filename, int x, int y, int r, long time, Color c) {
 		super(filename, x, y, r, time);
 		setColor(c);
